Separate out file storage functionality into its own class
